pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mimi"
version = "0.1.0"
authors = [
    {name = "MiMi Team", email = "example@example.com"},
]
description = "Multi-Agent Multi-Model Framework"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "fastapi>=0.100.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.13",
    "langchain-core>=0.1.15",
    "requests>=2.30.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/MiMi"
"Bug Tracker" = "https://github.com/yourusername/MiMi/issues" 